requirements.txt
azure-functions==1.20.0
azure-functions-durable==1.2.9
azure-storage-blob==12.23.0
azure-cosmos==4.8.0
azure-identity==1.19.0
azure-ai-documentintelligence==1.0.0b4
openai==1.54.4
httpx==0.27.0
pytest==8.3.3
pytest-asyncio==0.24.0
